Rework, add keyboard support
This is a fairly major (read: breaking) change. For now I've dropped the
ability to define a tree via static markup. Instead, you must pass a
node object into the root `ivy-tree` component. At a minimum, this node
must have a `children` property which is an array of its child nodes.

The `ivy-tree` component will recursively yield each node and each of
its children, and each of their children, and so on. Here's an example
of how to use it:

```handlebars
{{#ivy-tree node=myNode as |nodeOrDescendant|}}
  <span>{{nodeOrDescendant.name}}</span>
{{/ivy-tree}}
```

The above would output a `<span>` tag containing the name of each node.
If the node has children, it would also output a subtree which contained
a separate `<span>` for each name, and so on.