Push tree structure down into the model
This removes the TreeNode mixin from the components, in favor of keeping
the actual tree structure in a separate model. This removes our
dependence on Ember's render process, which has changed quite a bit
since this addon was first written. This also allows us to fix the
deprecation warnings (and later errors) related to setting a property
during `didInsertElement`.