Expose `buildTree` function
Previously, a POJO `model` property would be passed in to `{{ivy-tree}}`
and the component would then build an in-memory tree of nodes. Now the
responsibility for building the tree is pushed out to the caller, and a
`buildTree` function is exposed for this purpose. This is a bit nicer
since it makes the tree nodes accessible to the caller, in case they
want to, for instance, expand or collapse nodes.